LPS-145823 Replace base extend pattern with vanilla schema builders
