--- modules/apps/commerce/commerce-product-definitions-web/src/main/java/com/liferay/commerce/product/definitions/web/internal/frontend/data/set/view/table/BaseProductTableFDSView.java
@@ -1,58 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.commerce.product.definitions.web.internal.frontend.data.set.view.table;
-
-import com.liferay.frontend.data.set.view.table.BaseTableFDSView;
-import com.liferay.frontend.data.set.view.table.FDSTableSchema;
-import com.liferay.frontend.data.set.view.table.FDSTableSchemaBuilder;
-import com.liferay.frontend.data.set.view.table.FDSTableSchemaBuilderFactory;
-
-import java.util.Locale;
-
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Igor Beslic
- */
-public abstract class BaseProductTableFDSView extends BaseTableFDSView {
-
-	@Override
-	public FDSTableSchema getFDSTableSchema(Locale locale) {
-		FDSTableSchemaBuilder fdsTableSchemaBuilder =
-			fdsTableSchemaBuilderFactory.create();
-
-		fdsTableSchemaBuilder = addActionLinkFields(fdsTableSchemaBuilder);
-
-		fdsTableSchemaBuilder = addFields(fdsTableSchemaBuilder);
-
-		return fdsTableSchemaBuilder.build();
-	}
-
-	protected FDSTableSchemaBuilder addActionLinkFields(
-		FDSTableSchemaBuilder fdsTableSchemaBuilder) {
-
-		return fdsTableSchemaBuilder.add(
-			"name", "name",
-			fdsTableSchemaField -> fdsTableSchemaField.setContentRenderer(
-				"actionLink"));
-	}
-
-	protected abstract FDSTableSchemaBuilder addFields(
-		FDSTableSchemaBuilder fdsTableSchemaBuilder);
-
-	@Reference
-	protected FDSTableSchemaBuilderFactory fdsTableSchemaBuilderFactory;
-
-}
